src/vllm-sr/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vllm-sr"
version = "0.3.0"
description = "vLLM Semantic Router - Intelligent routing for Mixture-of-Models"
authors = [{name = "vLLM-SR Team"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
keywords = ["vllm", "semantic-router", "llm", "routing", "caching"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.1.7",
    "pyyaml>=6.0.2",
    "jinja2>=3.1.4",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "huggingface_hub[cli]>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "black>=22.0.0",
]

[project.scripts]
vllm-sr = "cli.main:main"

[project.urls]
Homepage = "https://github.com/vllm-project/vllm-semantic-router"
Documentation = "https://github.com/vllm-project/vllm-semantic-router/blob/main/README.md"
Repository = "https://github.com/vllm-project/vllm-semantic-router"
Issues = "https://github.com/vllm-project/vllm-semantic-router/issues"

[tool.setuptools]
packages = ["cli", "cli.commands", "cli.templates"]

[tool.setuptools.package-data]
"cli.templates" = ["*.yaml", "*.json", "*.ini"]

[tool.pytest.ini_options]
addopts = ["-v"]
testpaths = ["tests"]
